+++ AssetTaggerDatabase/dbo/Tables/Department/Department.sql
@@ -5,7 +5,7 @@
     [Id] UNIQUEIDENTIFIER CONSTRAINT [DF_Department_Id] DEFAULT (NEWID()) NOT NULL,
     CONSTRAINT [PK_Department] PRIMARY KEY NONCLUSTERED ([Id] ASC),
 
-    [RowNumber] INT IDENTITY (1, 1),
+    [RowNumber] BIGINT IDENTITY (1, 1),
     CONSTRAINT [AK_Department_RowNumber] UNIQUE CLUSTERED ([RowNumber] ASC),
 
     -- Non-nullable columns.
